+++ abis/hades/pid_t.h
@@ -2,6 +2,6 @@
 #ifndef _ABIBITS_PID_T_H
 #define _ABIBITS_PID_T_H
 
-typedef long int pid_t;
+typedef int pid_t;
 
 #endif // _ABIBITS_PID_T_H
